+++ src/Profiler/FeatureDataCollector.php
@@ -2,9 +2,9 @@
 
 namespace Flagception\Bundle\FlagceptionBundle\Profiler;
 
-use Flagception\Bundle\FlagceptionBundle\Bag\FeatureResultBag;
+use Flagception\Bundle\FlagceptionBundle\Activator\TraceableChainActivator;
 use Exception;
-use Flagception\Bundle\FlagceptionBundle\Model\ResultGroup;
+use Flagception\Decorator\ChainDecorator;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\DataCollector\DataCollector;
@@ -18,20 +18,29 @@ use Symfony\Component\HttpKernel\DataCollector\DataCollector;
 class FeatureDataCollector extends DataCollector
 {
     /**
-     * The feature result bag
+     * The profiler chain activator
      *
-     * @var FeatureResultBag
+     * @var TraceableChainActivator
      */
-    private $resultBag;
+    private $chainActivator;
+
+    /**
+     * The chain decorator
+     *
+     * @var ChainDecorator
+     */
+    private $chainDecorator;
 
     /**
      * FeatureDataCollector constructor.
      *
-     * @param FeatureResultBag $resultBag
+     * @param TraceableChainActivator $chainActivator
+     * @param ChainDecorator $chainDecorator
      */
-    public function __construct(FeatureResultBag $resultBag)
+    public function __construct(TraceableChainActivator $chainActivator, ChainDecorator $chainDecorator)
     {
-        $this->resultBag = $resultBag;
+        $this->chainActivator = $chainActivator;
+        $this->chainDecorator = $chainDecorator;
     }
 
     /**
@@ -39,52 +48,137 @@ class FeatureDataCollector extends DataCollector
      */
     public function collect(Request $request, Response $response, Exception $exception = null)
     {
-        $grouped = [];
+        $this->data = [
+            'summary' => [
+                'features' => 0,
+                'activeFeatures' => 0,
+                'inactiveFeatures' => 0,
+                'corruptFeatures' => 0
+            ],
+            'requests' => [],
+            'activators' => [],
+            'decorators' => [],
+            'trace' => $this->chainActivator->getTrace()
+        ];
 
-        foreach ($this->resultBag->all() as $result) {
-            if (array_key_exists($result->getFeatureName(), $grouped)) {
-                $group = $grouped[$result->getFeatureName()];
-            } else {
-                $group = new ResultGroup($result->getFeatureName());
-            }
+        // Activators
+        foreach ($this->chainActivator->getActivators() as $offset => $activator) {
+            $name = $activator->getName();
 
-            if ($result->isActive() && $activator = $result->getActivator()) {
-                $group->addActivator($activator);
-            }
-
-            if ($result->isActive()) {
-                $group->increaseActive();
-            } else {
-                $group->increaseInactive();
-            }
-
-            $grouped[$result->getFeatureName()] = $group;
+            $this->data['activators'][$name] = [
+                'priority' => ++$offset,
+                'name' => $name,
+                'requests' => 0,
+                'activeRequests' => 0,
+                'inactiveRequests' => 0,
+            ];
         }
 
-        $this->data = [
-            'results' => $this->resultBag->all(),
-            'groupedResults' => $grouped
-        ];
+        // Decorators
+        foreach ($this->chainDecorator->getDecorators() as $offset => $decorator) {
+            $name = $decorator->getName();
+
+            $this->data['decorators'][$name] = [
+                'priority' => ++$offset,
+                'name' => $name
+            ];
+        }
+
+        // Analyze trace
+        foreach ($this->chainActivator->getTrace() as $trace) {
+            if (!isset($this->data['requests'][$trace['feature']])) {
+                $this->data['requests'][$trace['feature']] = [
+                    'requests' => 0,
+                    'activeRequests' => 0,
+                    'inactiveRequests' => 0,
+                    'activators' => []
+                ];
+            }
+
+            $featureTrace = &$this->data['requests'][$trace['feature']];
+            $featureTrace['requests']++;
+
+            if ($trace['result'] === true) {
+                $featureTrace['activeRequests']++;
+            } else {
+                $featureTrace['inactiveRequests']++;
+            }
+
+            foreach ($trace['stack'] as $activator => $result) {
+                if ($result === true && !in_array($activator, $featureTrace['activators'], true)) {
+                    $featureTrace['activators'][] = $activator;
+                }
+
+                $this->data['activators'][$activator]['requests']++;
+
+                if ($result === true) {
+                    $this->data['activators'][$activator]['activeRequests']++;
+                } else {
+                    $this->data['activators'][$activator]['inactiveRequests']++;
+                }
+            }
+        }
+
+        $this->data['summary']['features'] = count($this->data['requests']);
+        foreach ($this->data['requests'] as $trace) {
+            if ($trace['activeRequests'] > 0 && $trace['inactiveRequests'] === 0) {
+                $this->data['summary']['activeFeatures']++;
+            } elseif ($trace['inactiveRequests'] > 0 && $trace['activeRequests'] === 0) {
+                $this->data['summary']['inactiveFeatures']++;
+            } else {
+                $this->data['summary']['corruptFeatures']++;
+            }
+        }
     }
 
     /**
      * Get all results
      *
-     * @return FeatureResultBag
+     * @return array
      */
-    public function getResults()
+    public function getRequests()
     {
-        return $this->data['results'];
+        return $this->data['requests'];
     }
 
     /**
-     * Get all grouped results / features
+     * Get all activators
      *
-     * @return ResultGroup[]
+     * @return array
      */
-    public function getGroupedResults()
+    public function getActivators()
     {
-        return $this->data['groupedResults'];
+        return $this->data['activators'];
+    }
+
+    /**
+     * Get all decorators
+     *
+     * @return array
+     */
+    public function getDecorators()
+    {
+        return $this->data['decorators'];
+    }
+
+    /**
+     * Get trace
+     *
+     * @return array
+     */
+    public function getTrace()
+    {
+        return $this->data['trace'];
+    }
+
+    /**
+     * Get summary
+     *
+     * @return array
+     */
+    public function getSummary()
+    {
+        return $this->data['summary'];
     }
 
     /**
@@ -100,6 +194,6 @@ class FeatureDataCollector extends DataCollector
      */
     public function reset()
     {
-        $this->resultBag->clear();
+        $this->data = [];
     }
 }